+++ jscomp/gentype/TranslateTypeExprFromTypes.ml
@@ -141,7 +141,7 @@ let translateConstr ~config ~paramsTranslation ~(path : Path.t) ~typeEnv =
             case 0 "Ok" paramTranslation1.type_;
             case 1 "Error" paramTranslation2.type_;
           ]
-        ~polymorphic:false ~unboxed:false
+        ~polymorphic:false ~unboxed:false ~customTag:None
     in
     {
       dependencies =
@@ -382,7 +382,7 @@ and translateTypeExprFromTypes_ ~config ~typeVarsGen ~typeEnv
       in
       let type_ =
         createVariant ~inherits:[] ~noPayloads ~payloads:[] ~polymorphic:true
-          ~unboxed:false
+          ~unboxed:false ~customTag:None
       in
       {dependencies = []; type_}
     | {noPayloads = []; payloads = [(_label, t)]; unknowns = []} ->
@@ -411,7 +411,7 @@ and translateTypeExprFromTypes_ ~config ~typeVarsGen ~typeEnv
       in
       let type_ =
         createVariant ~inherits:[] ~noPayloads ~payloads ~polymorphic:true
-          ~unboxed:false
+          ~unboxed:false ~customTag:None
       in
       let dependencies =
         payloadTranslations